initial commit
This is the first commit to this repo. It contains a basic TTS extension
for Firefox. It was hacked together in about two hours and I lost
interest working on it right now, because now it is all about figuring
out how to split and replace certain words to have a better experience.
It also is sufficient enough for my use case as it is right now.

The extension figures out itself if the language to speak is German
or English by simply looking for the 500 most used words of each language.