Splitting text at dots into substrings for better performance
It seems like, that Firefox has a buffer of 64 words when doing TTS.
Aber 64 words it shortly stops right in the center of a sentence and
makes the whole experience way less gread.

With this change the text ist splitted into senctences (at dots ".") and
adds all of them into the queue. This results into having longer breaks
between sentences, but it makes the whole experience way better when
reading longer texts.